remove selected project instead of current project
`m-x counsel-projectile-switch-project` `m-o r` removes the *current* project
from `projectile-known-projects` instead of the selected one.

The expected behaviour is for the action to operate on the *selection*, which is
the case for the rest of the actions.

I was not able to determine why this was happening. The use of dynamic binding
and my inexperience with the elisp debugger were a limitation; It looks like it
might be an inconsistency in projectile, but not necessarily a bug.