Improve counsel-projectile-find-file performance
Current code will show significant lag for projects with large numbers
of files.  Tests with 400K files shows 15-20s lag for
counsel-projectile-find-file.  Profiling showed that this was
due to the preprocessing of projectile-current-project-files to add
'ivy-virtual face to files not already in a buffer.  To fix this,
a display transformer was added for this command so that the face
modification was applied only to at most ivy-height files.

The preprocessed file list was also used by counsel-projectile
code paths, so some additional modifications were made to ensure
that that code path does not change behavior.  Any similar issues in
counsel-projectile performance was not addressed in this commit.

Details:

* move 'ivy-virtual face change to a display transformer function
* use projectile-current-project-files instead of counsel-projectile--file-list
  for ivy-read.
* reduce the scope of counsel-projectile--file-list, making it return
  only unvisited files, and renaming it accordingly.  This is now only
  used indirectly by the counsel-projectile function.
* change ivy-switch-buffer-transformer to propertize files at display time.