Advise ivy-read-action for MELPA Stable and ELPA users
For MELPA Stable and ELPA users, until the next version of ivy is
released, we need to advise `ivy-read-action' to implement
multi-character action keys. The advice only operates on
counsel-projectile commands (identified by the fact that
`ivy-state-caller' starts with \"counsel-projectile\"), so as not to
mess up with the general ivy setup.