counsel-projectile, cp-switch-to-buffer: fix default sorting
When the collection is a function, it is this function that must be
added to `ivy-sort-functions-alist`. Adding the caller instead has no
effect.

Note that at the moment, sorting may not work correctly if the default
is changed, due to an issue in ivy with sorting with collection
functions. See abo-abo/swiper/pull/1592.